Created working attacks! Also fixed up get methods for monsters and players in the tile class

+++ sketch_180523a/Monster.pde
@@ -1,6 +1,6 @@
 public class Monster extends Character{
     
-  public Monster(int h, int m, String n, int _x, int _y, char direction){
+  public Monster(float h, float m, String n, float _x, float _y, char direction){
     health = h;
     mana = m;
     name = n;
