+++ sketch_180523a/sketch_180523a.pde
@@ -7,16 +7,16 @@ public class PixelGungeon{
     }
 
     public void mapGen(){
-  for (int r = 0; r<map.length; r++){
-      for (int c = 0; c<map[0].length; c++){
+  for (float r = 0; r<map.length; r++){
+      for (float c = 0; c<map[0].length; c++){
     map[r][c] = new Tile(r , c , false);
       }
   }
-  for (int n=0; n<map[0].length; n++){
+  for (float n=0; n<map[0].length; n++){
       map[0][n] = new Tile(0 , n , true);
       map[map.length-1][n] = new Tile( map.length - 1 , n , true);
   }
-  for (int n=0; n<map.length; n++){
+  for (float n=0; n<map.length; n++){
       map[n][0] = new Tile(n , 0 , true) ;
       map[n][map[0].length-1] = new Tile( n  , map[0].length - 1 , true);
   }
@@ -36,8 +36,8 @@ public class PixelGungeon{
   if (map.length==0){
       return "[]";
   }
-  for (int c=0; c<map.length; c++){
-     for (int r=0; r<map[0].length; r++){
+  for (float c=0; c<map.length; c++){
+     for (float r=0; r<map[0].length; r++){
         if(map[r][c].isWall)
         {
         dump += "W";
@@ -63,8 +63,8 @@ public class PixelGungeon{
     public void moveRight() 
       {
         System.out.println("Move right");
-          int row = playerStore.getX();
-          int col = playerStore.getY();
+          float row = playerStore.getX();
+          float col = playerStore.getY();
           if(row + 1 < map.length && !(map[row+1][col].checkMonster()) && !(map[row+1][col].isWall()))
             {
                map[row][col].removePlayer();
@@ -77,8 +77,8 @@ public class PixelGungeon{
     public void moveLeft() 
       {
         System.out.println("Move left");
-          int row = playerStore.getX();
-          int col = playerStore.getY();
+          float row = playerStore.getX();
+          float col = playerStore.getY();
           if(row - 1 >= 0 && !(map[row-1][col].checkMonster()) && !(map[row-1][col].isWall()))
           {
               map[row][col].removePlayer();
@@ -91,8 +91,8 @@ public class PixelGungeon{
     public void moveUp()
       {
         System.out.println("Move up");
-          int row = playerStore.getX();
-          int col = playerStore.getY();
+          float row = playerStore.getX();
+          float col = playerStore.getY();
           if(col  - 1 >= 0 && !(map[row][col-1].checkMonster()) && !(map[row][col-1].isWall()))
             {
                 map[row][col].removePlayer();
@@ -104,8 +104,8 @@ public class PixelGungeon{
     public void moveDown() 
       {
           System.out.println("Move down");
-          int row = playerStore.getX();
-          int col = playerStore.getY();
+          float row = playerStore.getX();
+          float col = playerStore.getY();
           if(col + 1 < map.length && !(map[row][col+1].checkMonster()) && !(map[row][col+1].isWall()))
             {
                 map[row][col].removePlayer();
@@ -119,8 +119,8 @@ public class PixelGungeon{
       
     
     public void display(){
-      for (int r=0; r<map.length; r++){
-        for (int c=0; c<map[0].length; c++){
+      for (float r=0; r<map.length; r++){
+        for (float c=0; c<map[0].length; c++){
           if(map[r][c].isWall())
           {
             fill(150);
